+++ database/migrations/2022_05_13_081612_create_siswas_table.php
@@ -17,8 +17,8 @@ class CreatesiswasTable extends Migration
         Schema::create('siswas', function (Blueprint $table) {
             $table->id();
             $table->string('nama_siswa');
-            $table->float('nis');
-            $table->float('nisn');
+            $table->integer('nis');
+            $table->integer('nisn');
             $table->string('nama_ayah');
             $table->string('nama_ibu');
             $table->string('alamat');
